Разбил на модули data и main
